Index entries only need the first value in the child.
Having index entries hold both the first and last values in a child enables
some minor optimizations, but I think having the index be half the size
is going to be a much bigger win.

This updates the estimator based on that, and then the design document
based on its new estimates.  It changes details (and cuts index sizes in
half) without changing the design in any major way.

Signed-off-by: Ben Pfaff <blp@feldera.com>